RecordingActivity has navigation drawer
- previously navigation drawer was on SummaryActivity
- SummaryActivity, SpeakerStatsActivity deleted (no longer used)
- LIUM-specific navigation drawer menuitems also removed
  - show_mfc_file, show_uem_seg_file, etc...
- "Show .raw" and .wav menuitems also removed (used for early testing, no longer needed)
- drawer tests moved to RecordingActivityTest
- formatting fixes on splash screen (extra new lines)
