--- app/src/main/res/layout/activity_summary.xml
@@ -1,160 +1,0 @@
-<android.support.v4.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/drawer_layout"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <!-- The main content view -->
-    <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical">
-
-        <ScrollView
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="4">
-
-            <GridLayout
-                android:id="@+id/speaker_duration_grid"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:columnCount="3"
-                android:paddingTop="25dp"
-                android:useDefaultMargins="true">
-
-                <TextView
-                    android:id="@+id/speaker_name_title"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="@string/speaker_name_title"
-                    android:textStyle="bold" />
-
-                <TextView
-                    android:id="@+id/speaker_time_title"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_gravity="right"
-                    android:text="@string/speaker_time_title"
-                    android:textStyle="bold" />
-
-                <TextView
-                    android:id="@+id/speaker_barchart_title"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="@string/speaker_barchart_title"
-                    android:textStyle="bold" />
-
-            </GridLayout>
-        </ScrollView>
-
-        <GridLayout
-            android:layout_width="fill_parent"
-            android:layout_height="wrap_content"
-            android:columnCount="4"
-            android:paddingEnd="5dp">
-
-            <TextView
-                android:id="@+id/textview_duration_header"
-                android:layout_columnWeight="1"
-                android:layout_gravity="right"
-                android:text="@string/text_duration_header"
-                android:textStyle="bold" />
-
-            <TextView
-                android:id="@+id/textview_max_header"
-                android:layout_columnWeight="1"
-                android:layout_gravity="right"
-                android:text="@string/text_max_header"
-                android:textStyle="bold" />
-
-            <TextView
-                android:id="@+id/textview_min_header"
-                android:layout_columnWeight="1"
-                android:layout_gravity="right"
-                android:text="@string/text_min_header"
-                android:textStyle="bold" />
-
-            <TextView
-                android:id="@+id/textview_average_header"
-                android:layout_columnWeight="1"
-                android:layout_gravity="right"
-                android:text="@string/text_average_header"
-                android:textStyle="bold" />
-
-            <TextView
-                android:layout_columnWeight="1"
-                android:layout_gravity="right"
-                android:text="@string/hmmss" />
-
-            <TextView
-                android:layout_columnWeight="1"
-                android:layout_gravity="right"
-                android:text="@string/hmmsss" />
-
-            <TextView
-                android:layout_columnWeight="1"
-                android:layout_gravity="right"
-                android:text="@string/hmmsss" />
-
-            <TextView
-                android:layout_columnWeight="1"
-                android:layout_gravity="right"
-                android:text="@string/hmmsss" />
-
-            <TextView
-                android:id="@+id/textview_duration"
-                android:layout_columnWeight="1"
-                android:layout_gravity="right" />
-
-            <TextView
-                android:id="@+id/textview_max"
-                android:layout_columnWeight="1"
-                android:layout_gravity="right" />
-
-            <TextView
-                android:id="@+id/textview_min"
-                android:layout_columnWeight="1"
-                android:layout_gravity="right" />
-
-            <TextView
-                android:id="@+id/textview_average"
-                android:layout_columnWeight="1"
-                android:layout_gravity="right" />
-        </GridLayout>
-
-        <GridLayout
-            android:layout_width="fill_parent"
-            android:layout_height="wrap_content"
-            android:columnCount="2">
-
-            <Button
-                android:id="@+id/button_new_meeting"
-                android:layout_width="0dp"
-                android:layout_height="match_parent"
-                android:layout_columnWeight="1"
-                android:onClick="newMeeting"
-                android:text="@string/button_new_meeting" />
-
-            <Button
-                android:id="@+id/button_share"
-                android:layout_width="0dp"
-                android:layout_height="match_parent"
-                android:layout_columnWeight="1"
-                android:onClick="share"
-                android:text="@string/button_share" />
-
-        </GridLayout>
-
-    </LinearLayout>
-
-    <!-- The navigation drawer -->
-    <android.support.design.widget.NavigationView
-        android:id="@+id/left_drawer"
-        android:layout_width="wrap_content"
-        android:layout_height="match_parent"
-        android:layout_gravity="start"
-        android:choiceMode="singleChoice"
-        app:headerLayout="@layout/drawer_header"
-        app:menu="@menu/drawer" />
-</android.support.v4.widget.DrawerLayout>
