+++ app/src/main/res/menu/drawer.xml
@@ -10,30 +10,6 @@
             android:onClick="launchMainActivity"
             android:title="@string/launch_main_activity" />
         <item
-            android:id="@+id/show_raw"
-            android:onClick="showRawFile"
-            android:title="@string/show_raw_file" />
-        <item
-            android:id="@+id/show_wav"
-            android:onClick="showWavFile"
-            android:title="@string/show_wav_file" />
-        <item
-            android:id="@+id/show_mfc_file"
-            android:onClick="showMfcFile"
-            android:title="@string/show_mfc_file" />
-        <item
-            android:id="@+id/show_uem_seg_file"
-            android:onClick="showUemSegFile"
-            android:title="@string/show_uem_seg_file" />
-        <item
-            android:id="@+id/show_s_seg_file"
-            android:onClick="showSSegFile"
-            android:title="@string/show_s_seg_file" />
-        <item
-            android:id="@+id/show_l_seg_file"
-            android:onClick="showLSegFile"
-            android:title="@string/show_l_seg_file" />
-        <item
             android:id="@+id/launch_packed_circle_activity"
             android:onClick="launchPackedCircleActivity"
             android:title="@string/launch_packed_circles_activity" />
